moving controllers to new arch
