+++ test/adapters/controllers/test_change_password_controller.py
@@ -1,7 +1,7 @@
 import pytest
 
-from src.adapters.controllers.change_password_controller import ChangePasswordController
-from src.adapters.controllers.login_user_controller import LoginUserController
+from src.modules.change_password.app.change_password_controller import ChangePasswordController
+from src.modules.login_user.app.login_user_controller import LoginUserController
 from src.adapters.controllers.update_user_controller import UpdateUserController
 from src.adapters.helpers.http_models import HttpRequest
 from src.domain.entities.enums import ROLE, ACCESS_LEVEL
@@ -16,7 +16,8 @@ class Test_ChangePasswordController:
             'login': 75599469093
         })
 
-        changePasswordController = ChangePasswordController(UserRepositoryMock())
+        changePasswordController = ChangePasswordController(
+            UserRepositoryMock())
         response = await changePasswordController(request)
         assert response.status_code == 200
         assert response.body == {
@@ -31,7 +32,8 @@ class Test_ChangePasswordController:
             'message': ""
         })
 
-        changePasswordController = ChangePasswordController(UserRepositoryMock())
+        changePasswordController = ChangePasswordController(
+            UserRepositoryMock())
         response = await changePasswordController(request)
         assert response.status_code == 200
         assert response.body == {
@@ -45,7 +47,8 @@ class Test_ChangePasswordController:
             'login': 27550611033
         })
 
-        changePasswordController = ChangePasswordController(UserRepositoryMock())
+        changePasswordController = ChangePasswordController(
+            UserRepositoryMock())
         response = await changePasswordController(request)
         assert response.status_code == 400
         assert response.body == {
@@ -59,7 +62,8 @@ class Test_ChangePasswordController:
             'login': 1232569856910
         })
 
-        changePasswordController = ChangePasswordController(UserRepositoryMock())
+        changePasswordController = ChangePasswordController(
+            UserRepositoryMock())
         response = await changePasswordController(request)
         assert response.status_code == 400
         assert response.body == {
@@ -73,15 +77,11 @@ class Test_ChangePasswordController:
             'login': "teste@qualquercoisa.com"
         })
 
-        changePasswordController = ChangePasswordController(UserRepositoryMock())
+        changePasswordController = ChangePasswordController(
+            UserRepositoryMock())
         response = await changePasswordController(request)
         assert response.status_code == 400
         assert response.body == {
             'result': False,
             'message': 'User not found'
         }
-
-
-
-
-
