+++ test/adapters/controllers/test_confirm_user_creation_controller.py
@@ -1,9 +1,9 @@
 import pytest
 
-from src.adapters.controllers.change_password_controller import ChangePasswordController
+from src.modules.change_password.app.change_password_controller import ChangePasswordController
 from src.adapters.controllers.confirm_change_password_controller import ConfirmChangePasswordController
 from src.adapters.controllers.confirm_user_creation_controller import ConfirmUserCreationController
-from src.adapters.controllers.login_user_controller import LoginUserController
+from src.modules.login_user.app.login_user_controller import LoginUserController
 from src.adapters.controllers.update_user_controller import UpdateUserController
 from src.adapters.helpers.http_models import HttpRequest
 from src.domain.entities.enums import ROLE, ACCESS_LEVEL
@@ -20,7 +20,8 @@ class Test_ConfirmChangePasswordController:
         })
 
         repository = UserRepositoryMock()
-        confirmUserCreationController = ConfirmUserCreationController(repository)
+        confirmUserCreationController = ConfirmUserCreationController(
+            repository)
         response = await confirmUserCreationController(request)
         assert response.status_code == 200
         u = await repository.getUserByCpfRne('54134054052')
@@ -34,7 +35,8 @@ class Test_ConfirmChangePasswordController:
         })
 
         repository = UserRepositoryMock()
-        confirmUserCreationController = ConfirmUserCreationController(repository)
+        confirmUserCreationController = ConfirmUserCreationController(
+            repository)
         response = await confirmUserCreationController(request)
         assert response.status_code == 303
 
@@ -46,6 +48,7 @@ class Test_ConfirmChangePasswordController:
         })
 
         repository = UserRepositoryMock()
-        confirmUserCreationController = ConfirmUserCreationController(repository)
+        confirmUserCreationController = ConfirmUserCreationController(
+            repository)
         response = await confirmUserCreationController(request)
         assert response.status_code == 404
