+++ test/adapters/controllers/test_login_user_controller.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.adapters.controllers.login_user_controller import LoginUserController
+from src.modules.login_user.app.login_user_controller import LoginUserController
 from src.adapters.controllers.update_user_controller import UpdateUserController
 from src.adapters.helpers.http_models import HttpRequest
 from src.domain.entities.enums import ROLE, ACCESS_LEVEL
@@ -76,6 +76,3 @@ class Test_LoginUserController:
         loginUserController = LoginUserController(UserRepositoryMock())
         response = await loginUserController(request)
         assert response.status_code == 400
-
-
-
